+++ app/controllers/users_controller.rb
@@ -2,7 +2,8 @@ class UsersController < ApplicationController
   include SessionsHelper
 
   before_action :set_user, only: %i[show edit update destroy]
-  before_action :require_login
+  before_action :require_login, except: %w[new create]
+  before_action :require_logout, only: %w[new create]
 
   # GET /users
   # GET /users.json
@@ -14,27 +15,22 @@ class UsersController < ApplicationController
   # GET /users/1.json
   def show; end
 
-  # GET /users/new
   def new
     @user = User.new
+    render 'auth/signup'
   end
 
   # GET /users/1/edit
   def edit; end
 
-  # POST /users
-  # POST /users.json
   def create
     @user = User.new(user_params)
 
-    respond_to do |format|
-      if @user.save
-        format.html { redirect_to @user, notice: 'User was successfully created.' }
-        format.json { render :show, status: :created, location: @user }
-      else
-        format.html { render :new }
-        format.json { render json: @user.errors, status: :unprocessable_entity }
-      end
+    if @user.save
+      login @user
+      redirect_to root_path, notice: 'Successful registration.'
+    else
+      redirect_back fallback_location: signup_path, notice: 'Registration failed.'
     end
   end