+++ app/models/user.rb
@@ -1,7 +1,20 @@
 class User < ApplicationRecord
+  before_destroy :destroy_relations
+
   has_many :opinions, dependent: :destroy
   has_many :followings, foreign_key: :follower_id
   has_many :followers, class_name: :Following, foreign_key: :followed_id
+
+  mount_uploader :photo, PhotoUploader
+  mount_uploader :cover_image, CoverImageUploader
+
+  serialize :pohto, JSON
+  serialize :cover_image, JSON
+
+  validates :username, uniqueness: true, presence: true, length: { maximum: 100 }
+  validates :full_name, presence: true, length: { maximum: 100 }
+  # validate :photo_size
+  # validate :cover_size
 
   def follow(user)
     followings.create({ followed_id: user.id })
@@ -13,5 +26,13 @@ class User < ApplicationRecord
 
   def post(content)
     opinions.create({ text: content })
+  end
+
+  private
+
+  def destroy_relations
+    followers.destroy_all
+    followings.destroy_all
+    opinions.destroy_all
   end
 end